requirements.txt
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
typing_extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
PyYAML==6.0
psycopg2==2.9.2 
flask_ngrok==0.0.25
